Add files via upload
These insights are then used by teams across the business to launch a new marketing campaign decide on features to build for an app, track the success of the app by measuring user engagement and improve the experience altogether while helping the business grow. You are working with product team of Instagram and the product manager has asked you to provide insights on the questions asked by the management team. 

Marketing: The marketing team wants to launch some campaigns, and they need your need your help with following:
1. Rewarding Most Loyal Users: People who have been using the platform for the longest time. 

Task: Find the 5 oldest users of the instagram from the database provided. 

2. Remind Inactive Users to Start posting: By sending them promotional emails to post their 1st photo. 
Task: Find the users who have never posted a single photo on Instagram. 

3.Declaring Content Winner: The team started a contest and the user who gets the most likes on a single photo will win the content now they wish to declare the winner. 
Task: Identify the winner of the contest and provide their details to the team. 

4. Hashtag researching: A partner brand wants to know, which hashtags to use in the post to reach the most people on the platform. 
Task: Identify and suggest the top 5 most commonly used hashtag on the platform

5. Launch AD Campaign: the team wants to know, which day would be the best day to launch ADs. 
Task: What day of the week do most users register on? Provide insights on when to schedule an ad campaign.